End of section 7

+++ API/Entities/Product.cs
@@ -9,6 +9,6 @@ public class Product
   public required string PictureUrl { get; set; }
   public required string Type { get; set; }
   public required string Brand { get; set; }
-  public required double Price { get; set; }
+  public required long Price { get; set; }
   public int QuantityInStock { get; set; }
 }
